--- src/domains/contrato/container/contrato-create/index.ts
@@ -1,2 +1,0 @@
-export { default as ContratoCreate } from './ContratoCreate';
-export { default as ContratoCreateV2 } from './ContratoCreate.V2';
