+++ src/domains/contrato/container/contrato-crud/index.ts
@@ -1,0 +1,2 @@
+export { default as ContratoCreate } from './ContratoCreate';
+export { default as ContratoEdit } from './ContratoEdit';
