--- src/domains/contrato/container/contrato-edit/index.ts
@@ -1,1 +1,0 @@
-export { default as ContratoEdit } from './ContratoEdit';
